requirements.txt
absl-py==2.1.0
accelerate==1.2.1
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.3.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
distro==1.9.0
dm-tree==0.1.8
docstring_parser==0.16
duckduckgo_search==7.1.1
durationpy==0.9
etils==1.11.0
evaluate==0.4.3
faiss-cpu==1.9.0.post1
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gensim==4.3.3
google-auth==2.37.0
google-pasta==0.2.0
#googleapis-common-protos==1.66.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
immutabledict==4.2.1
importlib_resources==6.4.5
ipywidgets==8.1.5
isoduration==20.11.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kagglehub==0.3.6
keras==3.7.0
kiwisolver==1.4.7
kneed==0.8.5
kubernetes==31.0.0
lab==8.3
langchain==0.3.13
langchain-chroma==0.1.4
langchain-community==0.3.13
langchain-core==0.3.28
langchain-huggingface==0.1.1
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.6
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.9
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib==3.10.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mlxtend==0.23.3
#mmcv==1.7.2
#mmcv-full==1.7.2
mmdet==2.26.0
mmengine==0.10.5
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.4.2
nltk==3.9.1
notebook==7.3.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
#onnxruntime==1.20.1
openai>=1.58.1,<2.0.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
#opentelemetry-api==1.29.0
#opentelemetry-exporter-otlp-proto-common==1.29.0
#opentelemetry-exporter-otlp-proto-grpc==1.29.0
#opentelemetry-instrumentation==0.50b0
#opentelemetry-instrumentation-asgi==0.50b0
#opentelemetry-instrumentation-fastapi==0.50b0
#opentelemetry-proto==1.29.0
#opentelemetry-sdk==1.29.0
#opentelemetry-semantic-conventions==0.50b0
#opentelemetry-util-http==0.50b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.13
overrides==7.7.0
pandas==2.2.3
pandocfilters==1.5.1
peft==0.14.0
pillow==11.0.0
plotly==5.24.1
posthog==3.7.4
primp==0.9.2
prometheus_client==0.21.1
promise==2.3
propcache==0.2.1
#protobuf==3.20.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pyparsing==3.2.0
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.3.1
shapely==2.0.6
shellingham==1.5.4
simple-parsing==0.1.6
simplejson==3.19.3
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.16.1
termcolor==2.5.0
terminado==0.18.1
terminaltables==3.1.10
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.6.0
torchsummary==1.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.0
txt2tags==3.9
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.34.0
#uvloop==0.21.0
watchfiles==1.0.3
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.0
xgboost==2.1.3
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3